improve frisk() test, extern var .h file
