+++ src/diskfrisk.c
@@ -3,11 +3,15 @@
 -> Set up CI
 -> Write remaining unit tests
 -> Search for hidden files
+-> Store options into config file, so search settings can persist,
+   Create interactive options menu to set
+-> Write makefile
 
 */
 
 #include <ctype.h>
 #include <dirent.h>
+#include "extern.h"
 #include <limits.h>
 #include "prototypes.h"
 #include <stdio.h>
@@ -24,26 +28,8 @@
 #define PMATCH    "grep:"         // User command to search by pattern match
 #define PM_LEN    strlen(PMATCH)  // Pattern match command length
 
-struct option_flags {
-    int csens;       // Case-sensitive search
-    int grep;        // Search by pattern match
-    int home;        // Search the home directory/user files
-    int openf;       // Open first occurance of filename match
-    int perm;        // Display permission errors
-    int sys;         // Search all system files excluding home/user files
-} option;
-
-struct error_flags {
-    int no_fn;       // No filename
-    int bad_flag;    // Illegal flag
-} error;
-
-int test = 0;
-int found = 0;       // Number of results
-char *dname = ROOT;  // Set by default
-
-clock_t start, end;
-double t_elapsed;
+extern int found;
+extern char *dname;
 
 
 /* Collect user input, parse optional flags */
@@ -76,10 +62,10 @@ char *input(int argc, char *argv[])
                     error.bad_flag = 1;
                     break;
             }
+            
+    dname = ((option.home && !option.sys)? HOME: ROOT);
     if (argc != 1)
         error.no_fn = 1;
-    if (option.home && !option.sys)
-        dname = HOME;
     if (!strncmp(*argv, PMATCH, PM_LEN)) {
         option.grep = 1;
         *argv = (*argv + PM_LEN);
@@ -89,54 +75,6 @@ char *input(int argc, char *argv[])
     display_state(x, *argv);
 
     return (error.no_fn + error.bad_flag) ? NULL : *argv;
-}
-
-
-/* Output current state to user */
-void display_state(char c, char *fname)
-{
-    if (!test) {
-
-        // No errors, display initial state to user
-        if (c == 1)
-            initial_state(fname); 
-
-        // Search complete
-        else if (!c)
-            result_line(fname); 
-
-        else
-            display_err(c);
-    }
-}
-
-
-void initial_state(char *fname)
-{
-    printf("\n\nDISKFRISK -- VERSION 0.0.0\n\n\n");
-    if (option.sys || (!option.home && !option.sys))
-        printf("System directory is being frisked...\n");
-    if (option.home || (!option.home && !option.sys))
-        printf("Home directory is being frisked...\n");
-    printf("Searching for%s: %s\n\n", ((option.grep) ? " pattern" : ""), fname);
-}
-
-
-void result_line(char *fname)
-{
-    printf("\n%d result%s, in %.2f seconds.\n\n", found, 
-        ((found != 1) ? "s" : ""), t_elapsed);
-    if (found < 1)
-        printf("%s not found...\n\n", fname);
-}
-
-
-void display_err(char c)
-{
-    if (error.bad_flag)
-        printf("\nIllegal option '%c'\n\n", c);
-    else if (error.no_fn)
-        printf("\nUsage: frisk -h -s <filename>\n\n");
 }
