+++ src/extern.h
@@ -1,0 +1,28 @@
+#ifndef EXTERN_H
+#define EXTERN_H
+
+#include "sysdep.h"
+#include <time.h>
+
+struct option_flags {
+    int csens;       // Case-sensitive search
+    int grep;        // Search by pattern match
+    int home;        // Search the home directory/user files
+    int openf;       // Open first occurance of filename match
+    int perm;        // Display permission errors
+    int sys;         // Search all system files excluding home/user files
+} option;
+
+struct error_flags {
+    int no_fn;       // No filename
+    int bad_flag;    // Illegal flag
+} error;
+
+int test;
+int found;           // Number of results
+char *dname;         // Set by default
+
+clock_t start, end;
+double t_elapsed;
+
+#endif
