+++ src/main.c
@@ -1,3 +1,4 @@
+#include "extern.h"
 #include "prototypes.h"
 #include <stdio.h>
 #include <stdlib.h>
